complexity-2 and minor refactoring

+++ src/service/itemService.ts
@@ -1,3 +1,7 @@
+import {
+  NotificationType,
+  receiverStatusCode,
+} from "../constants/appConstants";
 import { Item } from "../interface/Menu";
 import itemRepository from "../repository/itemRepository";
 import notificationService from "./notificationService";
@@ -5,13 +9,11 @@ class ItemService {
   async addNewItem(item: Item) {
     try {
       const result = await itemRepository.addNewItem(item);
-      const notificationDetails = {
-        message: `${item.name} has been added to menu at price ${item.price}`,
-        notificationType: 1,
-        receiverStatusCode: 3,
-      };
+      const notificationMessage = `${item.name} has been added to menu at price ${item.price}`;
       const notificationResult = await notificationService.saveNewNotification(
-        notificationDetails
+        notificationMessage,
+        NotificationType["New Item Added"],
+        receiverStatusCode.Both
       );
       return {
         sucess: true,
@@ -25,14 +27,22 @@ class ItemService {
 
   async updateItemPrice(item: Item) {
     try {
-      const result = await itemRepository.updatePrice(item);
-      return {
-        success: true,
-        message: "Price Updated Successfully!",
-        type: "message",
-      };
+      const count = await itemRepository.checkItemExists(item.name!);
+      if (count > 0) {
+        const result = await itemRepository.updatePrice(item);
+        return {
+          success: true,
+          message: "Price Updated Successfully!",
+          type: "message",
+        };
+      } else {
+        return {
+          success: false,
+          message: `Item with name '${name}' does not exist.`,
+          type: "error",
+        };
+      }
     } catch (error) {
-      console.log(error);
       return {
         success: false,
         message: error,
@@ -43,20 +53,22 @@ class ItemService {
 
   async updateItemAvailibilityStatus(item: Item) {
     try {
-      const result = await itemRepository.updateAvailibilityStatus(item);
-      const notificationDetails = {
-        message: `${item.name}'s availibility status has been changed`,
-        notificationType: 2,
-        receiverStatusCode: 2,
-      };
-      const notificationResult = await notificationService.saveNewNotification(
-        notificationDetails
-      );
-      return {
-        success: true,
-        message: "Availability Status Changed Successfully !",
-        type: "message",
-      };
+      const count = await itemRepository.checkItemExists(item.name!);
+      if (count > 0) {
+        const result = await itemRepository.updateAvailibilityStatus(item);
+        const notificationMessage = `${item.name}'s availibility status has been changed`;
+        const notificationResult =
+          await notificationService.saveNewNotification(
+            notificationMessage,
+            NotificationType["Availibility Status Changed"],
+            receiverStatusCode.Employee
+          );
+        return {
+          success: true,
+          message: "Availability Status Changed Successfully !",
+          type: "message",
+        };
+      }
     } catch (error) {
       return {
         success: false,
@@ -66,14 +78,23 @@ class ItemService {
     }
   }
 
-  async deleteItem(itemName: Item) {
+  async deleteItem(itemName: string) {
     try {
-      const result = await itemRepository.deleteItem(itemName);
-      return {
-        success: true,
-        message: "Item Deleted Successfully!",
-        type: "message",
-      };
+      const count = await itemRepository.checkItemExists(itemName);
+      if (count > 0) {
+        const result = await itemRepository.deleteItem(itemName);
+        return {
+          success: true,
+          message: "Item Deleted Successfully!",
+          type: "message",
+        };
+      } else {
+        return {
+          success: false,
+          message: "No such Item exist",
+          type: "message",
+        };
+      }
     } catch (error) {
       return {
         success: false,
